% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jama_survplot.R
\name{ggsurvplot_jama}
\alias{ggsurvplot_jama}
\title{ggsurvplot_jama.}
\usage{
ggsurvplot_jama(surv_fit, x_label = "Time", y_label)
}
\arguments{
\item{x_label}{the title of x axis.}

\item{y_label}{the title of y axis.}

\item{fit}{an object of class survfit.}
}
\description{
Making KM survival curves and risk table in JAMA publication-ready format.
}
